Adding authorization to update and delete blogs

+++ controllers/blogController.js
@@ -1,5 +1,6 @@
 const Blog = require("./../models/blogModel");
 const catchAsync = require("./../utils/catchAsync");
+const AppError = require("./../utils/AppError");
 
 exports.getAllBlogs = catchAsync(async (req, res) => {
   const blogs = await Blog.find();
@@ -18,8 +19,11 @@ exports.getBlog = catchAsync(async (req, res) => {
 });
 
 exports.createBlog = catchAsync(async (req, res) => {
-  console.log(req.body);
-  await Blog.create(req.body);
+  const author = req.user._id;
+  const { title, content, tags } = req.body;
+
+  await Blog.create({ title, content, tags, author });
+
   res.status(201).json({
     status: "success",
     message: "Blog created successfully!",
@@ -27,6 +31,18 @@ exports.createBlog = catchAsync(async (req, res) => {
 });
 
 exports.updateBlog = catchAsync(async (req, res, next) => {
+  const blog = await Blog.findById(req.params.id);
+
+  if (!blog) {
+    return next(new AppError("No blog found with that ID", 404));
+  }
+
+  if (blog.author.toString() !== req.user.id) {
+    return next(
+      new AppError("You do not have permission to update this blog", 403)
+    );
+  }
+
   await Blog.findByIdAndUpdate(req.params.id, req.body, {
     runValidators: true,
   });
@@ -37,8 +53,20 @@ exports.updateBlog = catchAsync(async (req, res, next) => {
   });
 });
 
-exports.deleteBlog = catchAsync(async (req, res) => {
-  const blog = await Blog.findByIdAndDelete(req.params.id);
+exports.deleteBlog = catchAsync(async (req, res, next) => {
+  const blog = await Blog.findById(req.params.id);
+
+  if (!blog) {
+    return next(new AppError("No blog found with that ID", 404));
+  }
+
+  if (blog.author.toString() !== req.user.id) {
+    return next(
+      new AppError("You do not have permission to delete this blog", 403)
+    );
+  }
+
+  await Blog.findByIdAndDelete(req.params.id);
   res.status(200).json({
     status: "success",
     message: "Blog deleted Successfully!",
